refactor(gui): extract reusable profile shell
