+++ crates/openlogi-desktop/src/features/mod.rs
@@ -6,4 +6,4 @@ pub mod keyboard;
 pub mod lighting;
 pub mod mouse;
 pub mod pointer;
-pub mod profile_scope;
+pub mod profiles;
